Finished global headers feature
Since multiple headers can have the same key (accordingly to here):
https://stackoverflow.com/a/4371395

I decided to remove the override feature for global headers, now they
are always added to the final headers in the request, except if the
"disable global headers" checkbox is ticked.

Also added a tooltip to the "disable global headers" checkbox.