Added back the feature to override global headers
This feature was added back because it seems the RFCs don't
seem to allow duplicated headers in the request, this way we
can use the duplicated as overriden mechanism (for the global).

Also some minor fixes related to the global headers feature and
made clear in global headers tooltips that they are global (in
the request table).